marquee: move to nuklear_console_marquee.h, rename to nk_console_marquee_tooltip_render
